v1.11
- Changes to the Prediction:
a. Reworked the PROdiciton Hitboxcalculation (missed some things in the
last update)
b. Added VPrediction Hitboxcaluclation
c. Added a faster method for the Collision to PROdiction
d. Tweaked some ranges
- Changes to the Predictionmenu:
a. New and cleaner look for the menu
b. Removed Hitboxcalculation for BoLPrediction
- Added a HealManagermenu
a. Toggle to use the HealManager for you and your team
b. Toggle to disable the use of heal for certain teammembers
c. Slider to determine the %-health that is needed for the heal
d. Added an option to prioritize yourself or the team for the heal
- Changes to the KillStealmenu
a. Added a toggle to enable/disable SmartKillSteal
- Changes to the Drawmenu:
a. Added a new Submenu "PermaShow"
b. Added a toggle for all permashows (enabled by default)
c. If you disable a permashow, you need to restart the script (2xF9)
d. Added the Collisionlinedrawing (disabled by default)
- Bugfixes:
a. Fixed SmartKillSteal - should! work correctly now
b. Fixed JumpSpotCircles - now will show even when LFC is enabled